+++ src/flair/flair_fusion.py
@@ -250,18 +250,33 @@ def detectfusions():
     samtools_filter_cmd = ('samtools', 'view', '-F', '2048', '-hb', '-')
     samtools_sort_cmd = ('samtools', 'sort', '-@', str(args.threads), '-o', args.output + '.syntheticAligned.bam', '-')
     samtools_index_cmd = ('samtools', 'index', args.output + '.syntheticAligned.bam')
-    # bamtobedcmd = ('bamToBed', '-bed12', '-i', args.output + '.syntheticAligned.bam')
     bamtobedcmd = ('bedtools', 'bamtobed', '-bed12', '-i', args.output + '.syntheticAligned.bam')
     getsscommand = ['python3', path + 'synthetic_splice_sites.py', args.output + '.syntheticAligned.bed',
                         args.output + '-syntheticReferenceAnno.gtf', args.output + '.syntheticAligned.SJ.bed', args.output + '-syntheticBreakpointLoc.bed', '8', '2', args.output + '-syntheticFusionGenome.fa']#'15', '2']
-    ##NOT ADDING GTF ANNOT TO correct or collapse - I think this will save time down the line
-    correctcommand = ['python3', path + 'flair_correct.py', '-t', args.threads, '-q', args.output + '.syntheticAligned.bed',
-                      '-f', args.output + '-syntheticReferenceAnno.gtf',
-                      '--output', args.output + '.syntheticAligned.flair', '--junction_bed', args.output + '.syntheticAligned.SJ.bed', '--ss_window', '8']
-    collapsecommand = ['python3', path + 'flair_collapse.py', '-t', args.threads, '-q', args.output + '.syntheticAligned.flair_all_corrected.bed',
-                      '-g', args.output + '-syntheticFusionGenome.fa', #'-f', args.output + '-syntheticReferenceAnno.gtf',
-                      '--output', args.output + '.syntheticAligned.flair', '-r', freadsname, '--end_window', '300', #'--stringent', '--check_splice', #'--annotation_reliant', 'generate',
-                      '--generate_map', '--quality', '0', '--support', '2', '--fusion_breakpoints', args.output + '-syntheticBreakpointLoc.bed', '--allow_paralogs']
+    use_transcriptome = False  # FIXME: tmp debugging
+    if use_transcriptome:
+        transcriptome_command = ['python3', path + 'flair_transcriptome.py',
+                                 '--genome_aligned_bam', args.output + '.syntheticAligned.bam',
+                                 '--genome', args.output + '-syntheticFusionGenome.fa',
+                                 '--gtf', args.output + '-syntheticReferenceAnno.gtf',
+                                 '--junction_bed', args.output + '.syntheticAligned.SJ.bed',
+                                 '--ss_window', '8',
+                                 '--generate_map',
+                                 '--quality', '0',
+                                 '--sjc_support', '2',
+                                 '--allow_paralogs',
+                                 '--end_window', '300',
+                                 '--fusion_breakpoints', args.output + '-syntheticBreakpointLoc.bed',
+                                 '--output', args.output + '.syntheticAligned.flair']
+    else:
+    ## NOT ADDING GTF ANNOT TO correct or collapse - I think this will save time down the line
+        correctcommand = ['python3', path + 'flair_correct.py', '-t', args.threads, '-q', args.output + '.syntheticAligned.bed',
+                          '-f', args.output + '-syntheticReferenceAnno.gtf',
+                          '--output', args.output + '.syntheticAligned.flair', '--junction_bed', args.output + '.syntheticAligned.SJ.bed', '--ss_window', '8']
+        collapsecommand = ['python3', path + 'flair_collapse.py', '-t', args.threads, '-q', args.output + '.syntheticAligned.flair_all_corrected.bed',
+                           '-g', args.output + '-syntheticFusionGenome.fa', #'-f', args.output + '-syntheticReferenceAnno.gtf',
+                           '--output', args.output + '.syntheticAligned.flair', '-r', freadsname, '--end_window', '300', #'--stringent', '--check_splice', #'--annotation_reliant', 'generate',
+                           '--generate_map', '--quality', '0', '--support', '2', '--fusion_breakpoints', args.output + '-syntheticBreakpointLoc.bed', '--allow_paralogs']
 
 
     ##currently need to run correct and collapse as subprocess because they expect specific args, need to fix this at some point I think
@@ -275,9 +290,13 @@ def detectfusions():
     pipettor.run([bamtobedcmd], stdout=args.output + '.syntheticAligned.bed')
     pipettor.run([getsscommand])
     print('done getting ss')
-    pipettor.run([correctcommand])
-    print('done with correct')
-    pipettor.run([collapsecommand])
+
+    if use_transcriptome:
+        pipettor.run(transcriptome_command)
+    else:
+        pipettor.run([correctcommand])
+        print('done with correct')
+        pipettor.run([collapsecommand])
 
     ##clean up isoform/gene names for args.output + '.combined.isoform.read.map.txt', '.isoforms.bed', '.isoforms.fa'
     oldnametonewname = {}